refactor: (#731) 멤버의 게시글 개수와 투표 개수 설정

+++ backend/src/main/java/com/votogether/domain/member/dto/response/MemberInfoResponse.java
@@ -15,9 +15,9 @@ public record MemberInfoResponse(
         Integer birthYear,
 
         @Schema(description = "작성한 게시글 수", example = "5")
-        int postCount,
+        long postCount,
 
         @Schema(description = "투표한 수", example = "10")
-        int voteCount
+        long voteCount
 ) {
 }
